Add files via upload
Added an "add ingredient" and "add step" buttons that will put a pop up window for user to add. 
Added a popup window to inform user before deletion is executed.
Added a filter for the search recipe page that will allow user to filter results by a tag.

We will have to do error checking and format everything correctly, but it appears to be functioning as wanted for final sprint (except for a full-on search bar, but maybe we will skip that).